+++ packages/components/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 - `Icon` loading behaviour to non-blocking, components using the `Icon` will no longer wait for it to load 
 
+#### Fixed
+- Box model of `Link Pure`
+- Focus of `Link Pure` with slotted anchor and hidden label
+
 ### [2.4.0-beta.1] - 2021-08-26
 
 #### Added